refactor(runtime-react): move framework-free helpers into src/core/helpers.ts
Re-port of the first strangler step onto the arora-9 provider
(docs/redesign/06-track-2-implementation.md §3.1, PR-1 scope): the ~40
module-scope pure helpers move verbatim out of VizijRuntimeProvider.tsx
into core/helpers.ts (no React, no render store), now including the
suspension-inference helpers (IMPLICIT_PAUSE_GAP_S / inferImplicitPause)
that PR #75 added to the loop policy.

The provider re-exports resolveConstraintDefault,
deriveProgramInputSeedValues, and mergeAssetBundle so existing importers
are unchanged. 72/72 tests pass; public surface unchanged.
